==> 01_ballotpedia_scaper_senate_cleanup.ipynb <==
import pandas as pd
import numpy as np

### Get senatorial candidates from wmp using the latest file

s = pd.read_csv("../datasets/candidates/wmpcand_012523_wmpid.csv")

s.columns

s.genelect_cd.value_counts(dropna=False)

s = s.loc[(s['genelect_cd']==1)&(s['cand_office']=='S')]

s.shape

s.head()

s = s[['wmpid', 'cand_name', 'full_name', 'cand_office_st', 'cand_party_affiliation']]

s.head()

len(s['wmpid'].unique())

### Load the data cleaned by by RA Jasmine

Note: 
- Jasmine's clean-up is based on a previous scraping result which is slightly different from ./data/bp2022_sen_scraped_face.csv
- wmpid was added to this file using both fuzzymatch and human coding

sen_c = pd.read_excel("./data/bp2022_sen_scraped_face_jasmine.xlsx")

sen_c.shape

sen_c.head(2)

Use Jasmine's urls when scraped bp_url is missing

sen_c['bp_face_url']= sen_c['bp_url']

sen_c['bp_face_url']=sen_c.apply(
    lambda row: row['hc_face_url'] if pd.isnull(row['bp_face_url']) else row['bp_face_url'],
    axis=1)

sen_c.head(2)

sen_c.loc[sen_c['bp_face_url'].isna()]

url=sen_c[['wmpid', 'bp_face_url']]

### Finalize face urls

df = s.merge(url, on='wmpid', how='left', indicator=True)

df._merge.value_counts()

check the missing one from Jasmine's file 
- No face image: https://ballotpedia.org/Walter_Sluzynsky

df.loc[df._merge=='left_only']

df.head()

df.loc[df.bp_face_url.isna()]

df['face_2022']='yes'

df['face_category']='senate_general'

df = df[['wmpid', 'bp_face_url', 'face_2022', 'face_category']]

df.shape

df.head()

df=df.rename(columns={'bp_face_url':'face_url'})

df.to_csv("./data/face_url_2022_senate_general.csv", index=True)
