==> 02_ballotpedia_scaper_house.ipynb <==
### Scrape House candidate face urls from ballotpedia

Note: 
- The output file will be different using the updated wmpcand_XXXXXX_wmpid.csv file
- Our RA Jasmine used the output of this script to clean up face urls
- DO NOT 'run all' this script, as it involves manual check for each state. 

import pandas as pd
import bs4 as bs # pulling data out of HTML and XML files.
import urllib.request # opening and reading URLs
import re # pattern matching

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

#Download the website's html
source = urllib.request.urlopen("https://ballotpedia.org/United_States_House_of_Representatives_elections,_2022").read()

#Parse the html
#This makes it look nice and structured for us, and much easier to read for the program
soup = bs.BeautifulSoup(source)

urls_css = "center small a"

urls = soup.select(urls_css)

extracted_urls = [url["href"] for url in urls]
len(extracted_urls)

full_urls = ["https://ballotpedia.org" + url_suffix for url_suffix in extracted_urls]

full_urls

#### Distinguish two types of html structure for different states

urls_1 = []
urls_2 = []

for i in full_urls:
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  res_css = "#District_1"
  res = soup.select(res_css)
  if len(res) == 0:
    urls_1.append(i)
  else:
    urls_2.append(i)

urls_1

len(urls_2)

### Load the master file

h = pd.read_csv("../datasets/candidates/wmpcand_101422_wmpid.csv")

h.columns

h.genelect_cd.value_counts(dropna=False)

h = h.loc[(h['genelect_cd']==1)&(h['cand_office']=='H')]

h.shape

h = h[['wmpid', 'cand_name', 'full_name', 'cand_office_st', 'cand_office_dist', 'cand_party_affiliation']]

h.shape

len(h.full_name.unique())

h.head()

### Scape house candidates' face urls and fuzzy match with candidates in the masterfile at the state level

#### 1. States with one layer

- Here is exmaple code for one state Wyoming
- For each state, manually check the matching results to see if they're ok
- Once finishing all six states, combine them 

d1 = pd.DataFrame(columns=['bp_name_raw', 'bp_url'])

for i in urls_1[5:6]:
  print(i)
  images = []
  bp_names = []
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  img_css = ".image-candidate-thumbnail"
  img = soup.select(img_css)
  img = img[0:5]
  images.extend(list([i['src'] for i in img]))
  n_css = "td.votebox-results-cell--text"
  nm = soup.select(n_css)
  nm = nm[0:5]
  bp_names.extend(list([url.text for url in nm]))
  dic = {'bp_name_raw':bp_names,'bp_url':images}
  df = pd.DataFrame(dic)
  d1=d1.append(df)

d1

df2 = d1['bp_name_raw'].str.rsplit("(", n=1, expand=True)

df2 = df2.rename(columns={0:'name'})

df3 =df2['name'].str.rsplit("(", n=1, expand=True)
df3 = df3.rename(columns={0:'name'})

df3['bp_name'] = df3['name'].str.replace("\n", "").replace("\t", "")

df3['bp_name']

df['bp_name']=df3['bp_name']

df['bp_url_filename'] = df['bp_url'].str.replace("https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/100/100/", '')

df['bp_name'] = df['bp_name'].str.strip()

df = df.drop_duplicates(subset=['bp_name'], keep='first')

df

WY = h.loc[h['cand_office_st']=='WY']

WY

def fuzzy_merge(df_1, df_2, key1, key2, threshold=90, limit=2):
    """
    :param df_1: the left table to join
    :param df_2: the right table to join
    :param key1: key column of the left table
    :param key2: key column of the right table
    :param threshold: how close the matches should be to return a match, based on Levenshtein distance
    :param limit: the amount of matches that will get returned, these are sorted high to low
    :return: dataframe with boths keys and matches
    """
    s = df_2[key2].tolist()
    
    m = df_1[key1].apply(lambda x: process.extract(x, s, limit=limit))    
    df_1['matches'] = m
    
    m2 = df_1['matches'].apply(lambda x: ', '.join([i[0] for i in x if i[1] >= threshold]))
    df_1['matches'] = m2
    
    return df_1

wy = fuzzy_merge(WY, df, 'full_name', 'bp_name', threshold=90)

wy.head()

wy= wy.merge(df, left_on='matches', right_on='bp_name', indicator=True, how='left')

wy.head()

Quick check if matches are ok

wy.loc[wy._merge=='both']

#### Combine six states' results


#res1 = pd.concat([ak, de, nd, sd, vt, wy])

#res1.shape

#res1

#### 2. States with two layers

- Here is exmaple code for one state New_Mexico
- For each state, manually check the matching results to see if they're ok
- Once finishing all 44 states, combine them 

images = []
bp_name_raw = []

for i in urls_2[28:29]:
  print(i)
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  dist_css = "dd dl dd i a"
  dist = soup.select(dist_css)
  # urls_css = "center small a"
  # urls = soup.select(urls_css)
  extracted_urls = [i["href"] for i in dist]
  st_urls = ["https://ballotpedia.org" + url_suffix for url_suffix in extracted_urls]
  print(st_urls)
  for i in st_urls:
    #if '27s' in i:
    if '27' in i:
        source2 = urllib.request.urlopen(i).read()
        soup2 = bs.BeautifulSoup(source2)
        img_css = ".image-candidate-thumbnail"
        img = soup2.select(img_css)
        img = img[0:5]
        #print(img)
        images.extend(list([i['src'] for i in img]))
        n_css = "td.votebox-results-cell--text"
        nm = soup2.select(n_css)
        nm = nm[0:5]
        #print(nm)
        bp_name_raw.extend(list([url.text for url in nm]))

len(images)

len(bp_name_raw)

h_dic = {'bp_name_raw':bp_name_raw,'bp_url':images}

df = pd.DataFrame(h_dic)

df2 = df['bp_name_raw'].str.rsplit("(", n=1, expand=True)

df2 = df2.rename(columns={0:'name'})

df3 =df2['name'].str.rsplit("(", n=1, expand=True)
df3 = df3.rename(columns={0:'name'})

df3['bp_name'] = df3['name'].str.replace("\n", "").replace("\t", "")

df3['bp_name']

df['bp_name']=df3['bp_name']

df['bp_url_filename'] = df['bp_url'].str.replace("https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/100/100/", '')

df['bp_name'] = df['bp_name'].str.strip()

df = df.drop_duplicates(subset=['bp_name'], keep='first')

df

NM = h.loc[h['cand_office_st']=='NM']

NM.shape

NM

nm = fuzzy_merge(NM, df, 'full_name', 'bp_name', threshold=90)

nm

nm = nm.merge(df, left_on='matches', right_on='bp_name', indicator=True, how='left')

nm.head()

nm._merge.value_counts()

nm.shape

#### Combine 44 states' results

# res2 = pd.concat([al, az, ar, ca, co, ct, fl, ga, hi, xid, il, xin, ia, ks, ky,
#                  la, me, md, ma, mi, mn, ms, mo, mt, ne, nv, nh, nj, nm, ny, nc, oh, ok,
#                  xor, pa, ri, sc, tn, tx, ut, va, wa, wv, wi
#                  ])

#res2.shape

#res2.head()

#res2.shape

### Combine all results

#res = pd.concat([res1, res2])

#res.head()

#res._merge.value_counts(dropna=False)

#res.to_csv("./data/bp2022_house_scraped_face.csv", index=False)
