==> 02_ballotpedia_scaper_house_cleanup.ipynb <==
import pandas as pd
import numpy as np

### Get house candidates from wmp 

h = pd.read_csv("../datasets/candidates/wmpcand_012523_wmpid.csv")

h.columns

h.genelect_cd.value_counts(dropna=False)

h = h.loc[(h['genelect_cd']==1)&(h['cand_office']=='H')]

h.shape

h.head()

h = h[['wmpid', 'cand_name', 'full_name', 'cand_office_st', 'cand_party_affiliation']]

h.head()

len(h['wmpid'].unique())

### Load the data cleaned by RA Jasmine

Note: 
- Jasmine's clean-up is based on a previous scraping result which is slightly different from ./data/bp2022_sen_scraped_face.csv
- wmpid was added to this file using both fuzzymatch and human coding

house_c = pd.read_excel("./data/bp2022_house_scraped_face_jasmine.xlsx")

house_c.shape

house_c.head(2)

Use Jasmine's urls when scraped bp_url is missing

house_c['bp_face_url']= house_c['bp_url']

house_c['bp_face_url']=house_c.apply(
    lambda row: row['hc_face_url'] if pd.isnull(row['bp_face_url']) else row['bp_face_url'],
    axis=1)

house_c.head(2)

len(house_c.loc[house_c['bp_url'].isna()])

house_c.loc[house_c['bp_face_url'].isna()]

url=house_c[['wmpid', 'bp_face_url']]

### Finalize face urls

df = h.merge(url, on='wmpid', how='left', indicator=True)

df._merge.value_counts()

Double check the missing ones from Jasmine's file

df.loc[df._merge=='left_only']

df.loc[df['wmpid']=='WMPID4631', 'bp_face_url']='https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/200/300/Chris-Bye.PNG'
df.loc[df['wmpid']=='WMPID2605', 'bp_face_url']='https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/200/300/TimLong.jpg'

df.loc[df._merge=='left_only']

Check all house candidates without face urls

df.loc[df.bp_face_url.isna()]

df['face_2022']='yes'

df['face_category']='house_general'

df = df[['wmpid', 'bp_face_url', 'face_2022', 'face_category']]

df=df.rename(columns={'bp_face_url':'face_url'})

df.shape

df.head()

df.to_csv("./data/face_url_2022_house_general.csv", index=True)
