==> 03_ballotpedia_scaper_gov.ipynb <==
### Scrape gubernatorial candidate face urls from ballotpedia

import pandas as pd
import bs4 as bs # pulling data out of HTML and XML files.
import urllib.request # opening and reading URLs
import re # pattern matching

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

#Download the website's html
source = urllib.request.urlopen("https://ballotpedia.org/Gubernatorial_elections,_2022").read()

#Parse the html
soup = bs.BeautifulSoup(source)

urls_css = "tbody tr a"

urls = soup.select(urls_css)

extracted_urls = [url["href"] for url in urls[5:185:5]]

len(extracted_urls)

extracted_urls[0]

extracted_urls.append('/South_Carolina_gubernatorial_election,_2022')

extracted_urls

full_urls = ["https://ballotpedia.org" + url_suffix for url_suffix in extracted_urls]

full_urls[1]

Now that we've got all the links, we need to visit each, extract what we need and save it somewhere.

#Now do this in a loop, for the first 5 urls
bp_urls = []
bp_names = []

for i in full_urls[:]:
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  img_css = "img.image-candidate-thumbnail"
  img = soup.select(img_css)
  img = img[0:5]
  bp_urls.extend(list([i['src'] for i in img]))
  n_css = "td.votebox-results-cell--text"
  nm = soup.select(n_css)
  nm = nm[0:5]
  bp_names.extend(list([url.text for url in nm]))

len(bp_urls)

len(bp_names)

df = pd.DataFrame(columns=('bp_name_raw', 'bp_url'))

df['bp_name_raw']=bp_names

df['bp_url']=bp_urls

df.tail()

df2 = df['bp_name_raw'].str.rsplit("(", n=1, expand=True)

df2 = df2.rename(columns={0:'name'})

df3 =df2['name'].str.rsplit("(", n=1, expand=True)
df3 = df3.rename(columns={0:'name'})

df3['bp_name'] = df3['name'].str.replace("\n", "").replace("\t", "")

df3['bp_name']

df['bp_name']=df3['bp_name']

df.head()

df.shape

df

df['bp_name'] = df['bp_name'].str.strip()

df = df.drop_duplicates(subset=['bp_name'], keep='first')

df.shape

### Load ABC gubernatorial candidate list 

g = pd.read_excel("./data/Priors 2022.xlsx")

g.head()

g.columns

Clean up this file

g = g[['State', 'Democrat','Republican', 'Other','Unnamed: 7', 'Unnamed: 9']]
g.head()

abc = pd.melt(g, id_vars=['State'], value_vars=['Democrat', 'Republican', 'Other','Unnamed: 7', 'Unnamed: 9'],
             value_name='abc_name_raw')

abc = abc.rename(columns={'State':'abc_state'})

abc = abc[['abc_name_raw', 'abc_state']]

abc = abc.loc[~abc['abc_name_raw'].isna()]

abc.shape

abc['abc_name'] = abc['abc_name_raw'].str.rsplit("(", n=1, expand=True)[0]

abc

abc.loc[abc['abc_name']=='None of these candidates']

abc = abc.drop([92])

abc.loc[abc['abc_name']=='None of these candidates']

### Fuzzy matching names

def fuzzy_merge(df_1, df_2, key1, key2, threshold=90, limit=2):
    """
    :param df_1: the left table to join
    :param df_2: the right table to join
    :param key1: key column of the left table
    :param key2: key column of the right table
    :param threshold: how close the matches should be to return a match, based on Levenshtein distance
    :param limit: the amount of matches that will get returned, these are sorted high to low
    :return: dataframe with boths keys and matches
    """
    s = df_2[key2].tolist()
    
    m = df_1[key1].apply(lambda x: process.extract(x, s, limit=limit))    
    df_1['matches'] = m
    
    m2 = df_1['matches'].apply(lambda x: ', '.join([i[0] for i in x if i[1] >= threshold]))
    df_1['matches'] = m2
    
    return df_1

dat = fuzzy_merge(abc, df, 'abc_name', 'bp_name', threshold=90)

dat

dat.loc[dat['matches'].isna()]

dat.shape

gov = dat.merge(df, left_on='matches', right_on='bp_name', indicator=True, how='left')

gov.head()

gov._merge.value_counts()

gov.loc[gov['_merge']=='left_only']

# Mannually correct this entry
gov.loc[gov['_merge']=='left_only', 'bp_url']='https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/100/100/MikeDeWine2015.jpg'
gov.loc[gov['_merge']=='left_only', 'bp_name']='Richard Michael DeWine'

gov.loc[gov['_merge']=='left_only']

gov.columns

gov = gov[['abc_name', 'abc_name_raw', 'abc_state', 'bp_name', 'bp_name_raw', 'bp_url']]

gov['bp_url_filename'] = gov['bp_url'].str.replace("https://s3.amazonaws.com/ballotpedia-api4/files/thumbs/100/100/", '')

gov.head(20)

gov.to_csv("./data/bp2022_gov_scraped_face_verified.csv", index=False)

gov.shape
