==> 03_ballotpedia_scaper_gov_cleanup.ipynb <==
import pandas as pd

gov = pd.read_csv("./data/bp2022_gov_scraped_face_verified.csv")

gov.head(2)

gov.shape

#### Load the master file

df = pd.read_csv("../datasets/people/person_2022.csv")

g = df.loc[df['gov2022_gencd']==1]

g.shape

g.head(2)

g.loc[g['full_name'].str.contains("Mills")]

# Laura used bp_name when adding them to person data
g = g.merge(gov, how='left', left_on='full_name', right_on='bp_name', indicator=True)

g._merge.value_counts()

g = g[['wmpid', 'bp_url']]

g['face_2022']='yes'

g['face_category']='gov_general'

g.shape

g = g.rename(columns={'bp_url':'face_url'})

g.head()

g.face_url.isna().sum()

g.to_csv("./data/face_url_2022_gov_general.csv", index=True)
