==> 04_incumbent_governor.ipynb <==
## Incumbent governor

import pandas as pd
import bs4 as bs # pulling data out of HTML and XML files.
import urllib.request # opening and reading URLs
import re # pattern matching

### Get states that were not up for election in 2022

st = pd.read_csv("./data/s_and_t.csv")

st.head()

st = st.loc[(st['fips_code']<60)&(st['fips_code']!=11)]

st.shape

Get states up for election in 2022

g = pd.read_csv("./data/bp2022_gov_scraped_face_verified.csv")

g.head()

len(g['abc_state'].unique())

# Seats up for election
s = list(g['abc_state'].unique())

sue = pd.DataFrame()
sue['state_name'] = s
sue['state_name'] = sue['state_name'].str.upper()

sue.head()

gov = st.merge(sue, on='state_name', how='left', indicator=True)

gov.head()

incum = gov.loc[gov['_merge']=='left_only']

incum.head()

incum.shape

incum.to_csv('./data/incumbent_governor_tocode.csv', index=False)

As https://www.nga.org/governors/ can't be scraped, manually add face urls and saved as 'incumbent_governor_2022.csv'
