==> 04_incumbent_governor_cleanup.ipynb <==
import pandas as pd

### Load face urls (manually created )

incum= pd.read_csv("./data/incumbent_governor_2022.csv")

incum.head(2)

incum.shape

### Load the master file

df = pd.read_csv("../datasets/people/person_2022.csv")

g = df.loc[df['gov2022_noelect']==1]

g.shape

g.head(2)

inc = g.merge(incum, how='left', on='full_name', indicator=True)

inc._merge.value_counts()

inc.head()

inc = inc[['wmpid', 'face_url']]

inc['face_2022']='yes'

inc['face_category']='incumbent_governor'

inc.shape

inc.head()

inc.shape

inc.to_csv("./data/face_url_2022_incumb_gov_general.csv", index=False)
