==> 05_other_scrapers.ipynb <==
## Other scrapers

Notes: Be careful when re-run this script as many face images are replaced with current office holder 
on relevant official websites

import pandas as pd
import bs4 as bs 
import urllib.request
import requests
import re 

Load wmpids 

person = pd.read_csv("../datasets/people/person_2022.csv")

### International leaders

#### G20 

The following code is simple but scrape smaller thumb images; instead, we use larger profile
images from their personal pages for better face recognition results

# url = 'https://en.wikipedia.org/wiki/G20'
# source = req.urlopen(url).read()
# soup = bs.BeautifulSoup(source)

# image_css = '.gallerybox .image img'
# images = soup.select(image_css)

# images_src = [i['src'][2:] for i in images]
# print(images_src)

# df = pd.DataFrame()
# df['url'] = images_src

# df.head()

#Download the website's html
#The English page does not have the pictures anymore, you can try using the French page: https://fr.wikipedia.org/wiki/Groupe_des_vingt
source = urllib.request.urlopen("https://en.wikipedia.org/wiki/G20").read()
soup = bs.BeautifulSoup(source)
urls_css = ".center p a"
urls = soup.select(urls_css)

l1 = urls[1:14:3]
l2 = urls[17::3]
leaders = l1+l2

l_names = [url.text for url in leaders]

extracted_urls = [url["href"] for url in urls]

w1 = extracted_urls[1:14:3]
w2 = extracted_urls[17::3]
pages = w1+w2

page_urls = ["https://en.wikipedia.org/" + url_suffix for url_suffix in pages]

g20 = []

for i in page_urls[:]:
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  img_css = ".infobox-image a img"  
  img = soup.select(img_css)
  g20.extend(list(set([i['src'] for i in img])))

g20_face_urls = ["https:" + url_suffix for url_suffix in g20]

len(g20_face_urls)

t = pd.DataFrame(g20_face_urls)

t = t.loc[t[0].str.contains("220px")]

l_urls = t[0].to_list()

len(l_urls)

l_urls.pop(-1)

l_urls.pop(16)

g20 = pd.DataFrame()
g20['face_name']=l_names
g20['face_url']=l_urls

# To make name consistent with names in the person file:
g20.loc[g20['face_name']=='Salman', 'face_name'] = 'Mohammed bin Salman'
# Replace the president of Brazil
g20.loc[g20['face_name']=='Luiz Inácio Lula da Silva', 'face_name'] = 'Jair Bolsonaro'
g20.loc[g20['face_name']=='Luiz Inácio Lula da Silva', 'face_url'] = 'https://upload.wikimedia.org/wikipedia/commons/thumb/8/8a/Jair_Bolsonaro_em_24_de_abril_de_2019_%281%29_%28cropped%29.jpg/220px-Jair_Bolsonaro_em_24_de_abril_de_2019_%281%29_%28cropped%29.jpg'

g20.shape

l1 = {'face_name': 'António Guterres', 'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/d/d1/Ant%C3%B3nio_Guterres_2021.jpg/220px-Ant%C3%B3nio_Guterres_2021.jpg"}
il = g20.append(l1, ignore_index = True)

il['face_category']  = 'IntLeader'

il.shape

Get wmp ids

wmp_il = person.loc[person['intl_leaders']==1]

wmp_il.shape

wmp_il = wmp_il[['full_name', 'wmpid']]

il = il.merge(wmp_il, how='left', left_on='face_name', right_on='full_name', indicator=True)

il.loc[il['face_name']=='Joe Biden', 'wmpid']='WMPID1289'

il.loc[il._merge =='left_only']

il

il = il[['wmpid', 'face_url', 'face_category']]

il.shape

il.head(2)

### Former presidents

res = requests.get("https://en.wikipedia.org/wiki/List_of_presidents_of_the_United_States")
soup = bs.BeautifulSoup(res.content, "html.parser")

names =[]

for name in soup.find_all('b')[0:46]:
    names.append(name.get_text())

urls=[]
for link in soup.find_all('img')[6:52]:
    urls.append(link.get('src'))

urls = ['https:' + str(i).replace('160px', '320px') for i in urls]

fp = pd.DataFrame(columns=('face_name', 'face_url'))

fp['face_name']=names

fp['face_url']=urls

fp.shape

fp.head()

wmp_fp= person.loc[person['former_uspres']==1]
wmp_fp =wmp_fp[['wmpid', 'full_name']]

fp = fp.merge(wmp_fp, left_on='face_name', right_on='full_name', how='left')

fp.loc[fp.wmpid.isna()]

wmp_fp.loc[wmp_fp['full_name'].str.contains("Carter")]

fp.loc[fp['face_name']=='James A. Garfield', 'wmpid'] ='WMPID5184'
fp.loc[fp['face_name']=='George W. Bush', 'wmpid'] ='WMPID5'
fp.loc[fp['face_name']=='Barack Obama', 'wmpid'] ='WMPID3'
fp.loc[fp['face_name']=='Richard Nixon', 'wmpid'] ='WMPID5196'
fp.loc[fp['face_name']=='Gerald Ford', 'wmpid'] ='WMPID5197'
fp.loc[fp['face_name']=='Jimmy Carter', 'wmpid'] ='WMPID5180'

fp['face_category']='FormerPres'

Drop duplicated Grover Cleveland

fp = fp.drop_duplicates(keep='last')

fp = fp[['wmpid', 'face_url', 'face_category']]

fp.head(2)

fp.shape

### Supreme Court Justices

All justices who took judicial oath after 1980

source = urllib.request.urlopen("https://www.supremecourt.gov/about/biographies.aspx").read()
soup = bs.BeautifulSoup(source)

urls_css = ".media-object"

urls = soup.select(urls_css)

names = [url["alt"] for url in urls]

urls = [url["src"] for url in urls]

urls = ["https://www.supremecourt.gov/about/" + url_suffix for url_suffix in urls]

s = pd.DataFrame()
s['face_name'] = names
s['face_url']=urls

n = s['face_name'].str.split(pat = ',', expand = True)

s['face_name'] = n[0]

s2 = {'face_name': 'Ruth Bader Ginsburg', 'face_url': "https://www.supremecourt.gov/about/justice_pictures/Official_Justice_Ginsburg_sm.jpg"}
s = s.append(s2, ignore_index = True)

s3 = {'face_name': 'William Rehnquist', 
      'face_url':'https://upload.wikimedia.org/wikipedia/commons/thumb/3/31/William_Rehnquist.jpg/220px-William_Rehnquist.jpg'}
s = s.append(s3, ignore_index = True)

s4 = {'face_name': 'Antonin Scalia', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/c/c5/Antonin_Scalia_Official_SCOTUS_Portrait.jpg/220px-Antonin_Scalia_Official_SCOTUS_Portrait.jpg"}
s = s.append(s4, ignore_index = True)

s.shape

s.loc[s['face_name']=="Sandra Day O'Connor (Retired)", 'face_name']="Sandra Day O'Connor"

Get wmpid

wmp_s = person[['full_name', 'wmpid']]

s = s.merge(wmp_s, left_on='face_name', right_on='full_name', how='left')

s.loc[s['face_name']=="John G. Roberts", 'wmpid'] ='WMPID10'
s.loc[s['face_name']=="Samuel A. Alito", 'wmpid'] ='WMPID9'
s.loc[s['face_name']=="Neil M. Gorsuch", 'wmpid'] ='WMPID15'
s.loc[s['face_name']=="Brett M. Kavanaugh", 'wmpid'] ='WMPID16'
s.loc[s['face_name']=="Amy Coney Barrett", 'wmpid'] ='WMPID17'
s.loc[s['face_name']=="Stephen G. Breyer", 'wmpid'] ='WMPID12'
s.loc[s['face_name']=="Ruth Bader Ginsburg", 'wmpid'] ='WMPID18'

s['face_category']='SupremeCourt'

s

s = s[['wmpid', 'face_url', 'face_category']]

s.head(2)

### Cabinet

source = urllib.request.urlopen("https://www.whitehouse.gov/administration/cabinet/").read()
soup = bs.BeautifulSoup(source)

urls_css = ".persongrid__item__inner__content .h4alt"

urls = soup.select(urls_css)

c_names = [url.text for url in urls]

cab_names = []
for sub in c_names:
    cab_names.append(sub.replace("\n", "").replace("\t", "").replace("Dr. ", "")), 

urls_css = ".persongrid__item__inner__image>div img"

urls = soup.select(urls_css)

urls = [url["src"] for url in urls]

len(urls)

c = pd.DataFrame()
c['face_name'] = cab_names
c['face_url']=urls

# Replace chief of staff Jess Zients with his predecessor Ron Klain in 2022
c.loc[c['face_name']=='Jeff Zients', 'face_name']='Ron Klain'
c.loc[c['face_name']=='Jeff Zients', 'face_url']='https://upload.wikimedia.org/wikipedia/commons/thumb/e/e8/Ron_Klain%2C_White_House_Chief_of_Staff.jpg/220px-Ron_Klain%2C_White_House_Chief_of_Staff.jpg'

Get wmpid

wmp = person[['full_name', 'wmpid']]

c = c.merge(wmp, left_on='face_name', right_on='full_name', how='left')

c.loc[c['wmpid'].isna()]

c.loc[c['face_name']=="Deb Haaland", 'wmpid'] ='WMPID1183'
c.loc[c['face_name']=="Marcia Fudge", 'wmpid'] ='WMPID1206'
c.loc[c['face_name']=="Shalanda Young ", 'wmpid'] ='WMPID5185'

c['face_category']='Cabinet'

c

c = c[['wmpid', 'face_url', 'face_category']]

c.head(2)

### Public health related

ph = pd.DataFrame()

ph1 = {'face_name': 'Anthony Fauci', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/6/61/Anthony_Fauci_2020.jpg/220px-Anthony_Fauci_2020.jpg"}
ph = ph.append(ph1, ignore_index = True)

ph2 = {'face_name': 'Tedros Adhanom Ghebreyesus', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/0/0b/Mukhisa_Kituyi%2C_Houlin_Zhao%2C_Tedros_Adhanom_Ghebreyesus_with_Sophia_-_AI_for_Good_Global_Summit_2018_%2841223188035%29_%28cropped%29.jpg/220px-Mukhisa_Kituyi%2C_Houlin_Zhao%2C_Tedros_Adhanom_Ghebreyesus_with_Sophia_-_AI_for_Good_Global_Summit_2018_%2841223188035%29_%28cropped%29.jpg"}
ph = ph.append(ph2, ignore_index = True)

#The Secretary of the Department of Health and Human Services
ph3 = {'face_name': 'Rachel L. Levine', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/2/28/Admiral_Rachel_L._Levine.jpg/220px-Admiral_Rachel_L._Levine.jpg"}
ph = ph.append(ph3, ignore_index = True)

#The Surgeon General
ph4 = {'face_name': 'Vivek H. Murthy', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/7/7b/Vivek_Murthy%2C_Surgeon_General_%28profile%29.jpg/220px-Vivek_Murthy%2C_Surgeon_General_%28profile%29.jpg"}
ph = ph.append(ph4, ignore_index = True)

# Director of CDC
ph5 = {'face_name': 'Rochelle Walensky', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/b/b9/Rochelle_Walensky%2C_CDC_Director_%28cropped%29.jpg/220px-Rochelle_Walensky%2C_CDC_Director_%28cropped%29.jpg"}
ph = ph.append(ph5, ignore_index = True)

# Director of NIH
ph6 = {'face_name': 'Lawrence A. Tabak', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/a/a3/Lawrence_A._Tabak_official_portrait.jpg/220px-Lawrence_A._Tabak_official_portrait.jpg"}
ph = ph.append(ph6, ignore_index = True)

# Director of FDA
ph7 = {'face_name': 'Robert Califf', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/5/51/Robert_M._Califf%2C_FDA_Commissioner.jpg/220px-Robert_M._Califf%2C_FDA_Commissioner.jpg"}
ph = ph.append(ph7, ignore_index = True)

ph['face_category']='PublicHealth'

ph

ph.loc[ph['face_name']=='Tedros Adhanom Ghebreyesus', 'face_category']='IntLeader'

get wmpid

ph = ph.merge(wmp, left_on='face_name', right_on='full_name', how='left')

ph

ph = ph[['wmpid', 'face_url', 'face_category']]

ph.head(2)

### Historical figures

h = pd.DataFrame()

h1 = {'face_name': 'Martin Luther King Jr.', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/0/05/Martin_Luther_King%2C_Jr..jpg/220px-Martin_Luther_King%2C_Jr..jpg",
     'face_category': 'HistFigure'}
h = h.append(h1, ignore_index = True)

wmp.loc[wmp['full_name'].str.contains('Martin Luther')]

h['wmpid']='WMPID5221'

h

h = h[['wmpid', 'face_url', 'face_category']]

h.head()

### Combine all results

df = pd.concat([il, fp, s, c, ph, h])

df.shape

df.head()

df['face_category'].value_counts()

df.to_csv("./data/other_face_urls_2022.csv", index=False)
