==> 06_sitting_senator_scrapers.ipynb <==
import pandas as pd
import bs4 as bs 
import urllib.request
import re 

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

#Download the website's html
source = urllib.request.urlopen("https://en.wikipedia.org/wiki/List_of_United_States_senators_in_the_117th_Congress").read()
soup = bs.BeautifulSoup(source)

urls_css = ".fn a"
urls = soup.select(urls_css)

names = [url.text for url in urls]
sites = [url["href"] for url in urls]
page_urls = ["https://en.wikipedia.org/" + site for site in sites]

len(page_urls)

s =[]

for index, i in enumerate(page_urls[:]):
  source = urllib.request.urlopen(i).read()
  soup = bs.BeautifulSoup(source)
  img_css = ".infobox-image a img"  
  img = soup.select(img_css)
  s.extend(list(set([i['src'] for i in img])))

u = ["https:" + item for item in s if (('logo' in item)==False)&(('250px' in item)==False)]

df = pd.DataFrame(columns=['face_name', 'face_url'])

df['face_name']=names
df['face_url']=u

df.head()

df.loc[df['face_name'].str.contains('Casey'), 'face_url'][25]

### Get masterfile wmpids

person = pd.read_csv("../datasets/people/person_2022.csv")

s_sen = person.loc[person['currsen_2022']==1]

s_sen = s_sen[['full_name', 'wmpid']]
s_sen.head()

def fuzzy_merge(df_1, df_2, key1, key2, threshold=90, limit=2):
    """
    :param df_1: the left table to join
    :param df_2: the right table to join
    :param key1: key column of the left table
    :param key2: key column of the right table
    :param threshold: how close the matches should be to return a match, based on Levenshtein distance
    :param limit: the amount of matches that will get returned, these are sorted high to low
    :return: dataframe with boths keys and matches
    """
    s = df_2[key2].tolist()
    
    m = df_1[key1].apply(lambda x: process.extract(x, s, limit=limit))    
    df_1['matches'] = m
    
    m2 = df_1['matches'].apply(lambda x: ', '.join([i[0] for i in x if i[1] >= threshold]))
    df_1['matches'] = m2
    
    return df_1

match = fuzzy_merge(s_sen, df, 'full_name', 'face_name', threshold=70)

match['matches'].isna().sum()

# Get false positives in mutliple matches
match['face_name'] = match['matches'].str.split(pat = ',', expand = True)[0]

match.head()

match = match.merge(df, how='left', on='face_name')

match.head()

# split error for face name Bob Casey, Jr.
match.loc[match['face_url'].isna()]

match.loc[match['full_name']=='Robert Casey', 'face_url']=df.loc[df['face_name'].str.contains('Casey'), 'face_url'][25]

ss = match[['wmpid', 'face_url']]

ss['face_category'] = 'SitSen'

ss.head()

ss.shape

ss.to_csv('./data/face_url_2022_sitting_senate.csv', index=False)
