==> 07_face_url_final_assemble.ipynb <==
import pandas as pd

### Legislative

#### Senatorial candidates in general election

sen = pd.read_csv("./data/face_url_2022_senate_general.csv")
sen = sen[['wmpid', 'face_url', 'face_category']]
sen.loc[sen['face_category']=='senate_general','face_category']='Sen'

sen.head()

sen.shape

#### House candidates in general election

house = pd.read_csv("./data/face_url_2022_house_general.csv")
house = house[['wmpid', 'face_url', 'face_category']]
house.loc[house['face_category']=='house_general','face_category']='House'

house.head()

house.shape

#### Sitting Senators in 2022

ss = pd.read_csv("./data/face_url_2022_sitting_senate.csv")

ss.head()

ss.shape

### Executive

#### Former presidents

Biden is included; it's face_category will be changed to 'Pres' in the end

other = pd.read_csv("./data/other_face_urls_2022.csv")

pres = other.loc[other['face_category']=='FormerPres']

pres.head()

pres.shape

#### Cabinet

cab = other.loc[other['face_category']=='Cabinet']

cab.shape

#### Gubernatorial candidates 

g = pd.read_csv("./data/face_url_2022_gov_general.csv")

g.shape

#### Sitting governors

sg = pd.read_csv("./data/face_url_2022_incumb_gov_general.csv")

sg.shape

#### Public health related

ph = other.loc[other['face_category']=='PublicHealth']

ph.shape

### Judicial

court = other.loc[other['face_category']=='SupremeCourt']

court.shape

### Other prominent politicians

pp = pd.DataFrame(columns=['wmpid', 'face_url', 'face_category'])

# Hillary Clinton
pp1 = {'wmpid': 'WMPID7', 
      'face_url': "https://upload.wikimedia.org/wikipedia/commons/thumb/e/ec/Hillary_Clinton_by_Gage_Skidmore_4_%28cropped%29.jpg/220px-Hillary_Clinton_by_Gage_Skidmore_4_%28cropped%29.jpg",
    'face_category':'PromPoli'}
pp = pp.append(pp1, ignore_index = True)

# Mike Pence
pp2 = {'wmpid': 'WMPID1', 
      'face_url': 'https://upload.wikimedia.org/wikipedia/commons/thumb/b/b9/Mike_Pence_official_Vice_Presidential_portrait.jpg/220px-Mike_Pence_official_Vice_Presidential_portrait.jpg',
    'face_category':'PromPoli'}
pp = pp.append(pp2, ignore_index = True)

#Robert Mueller
pp3 = {'wmpid': 'WMPID8', 
      'face_url': 'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e1/Director_Robert_S._Mueller-_III.jpg/220px-Director_Robert_S._Mueller-_III.jpg',
    'face_category':'PromPoli'}
pp = pp.append(pp3, ignore_index = True)

pp.shape

pp.head()

#### International leaders

Biden is included; it's face_category will be changed to 'Pres' in the end

il = other.loc[other['face_category']=='IntLeader']

il.shape

#### Historical figures

h = other.loc[other['face_category']=='HistFigure']

h.shape

### Assemble all faces

d = pd.concat([sen, house, ss, pres, cab, g, sg, ph, court, pp, il, h])

d = d[['wmpid', 'face_url', 'face_category','face_2022']]

d.head()

d['face_2022']='yes'

d = d.rename(columns={'face_url':'face_url_2022', 'face_category':'face_category_2022'})

d.head()

d.loc[d['wmpid']=='WMPID1289', 'face_category_2022'] ='Pres'

person = pd.read_csv("../datasets/people/person_2022.csv", usecols =['wmpid', 'full_name'])

d= d.merge(person, on='wmpid', how='left')

Check duplicates

#check = d.loc[d.duplicated(subset=['wmpid'], keep=False)].sort_values(by=['wmpid'])
#check.head(20)

d.shape

d.head()

d.to_csv("./data/result_face_url_2022_raw.csv", index=False)
